Various changes to make clients react more on state from server.
Had an issue where if a player joined later they would not see the state
of the world as it had progressed since the game began. State for if a
door is opened or not is now kept on the server and when a new player
joins they get the state of the "world" and can render a correct
version. Possible issues is if a player looses connection for a while
and things happen during that time. They would not be received by the
client and this is a major issue. The best solution i can think of right
now is to keep all state on the server and emit a state every x ms that
is the truth that the clients will have to handle how to react to...